changed price to float field

+++ hospital/models.py
@@ -9,7 +9,7 @@ class Department(models.Model):
 
 class Medicine(models.Model):
     name = models.CharField(max_length=100)
-    price = models.IntegerField(blank=True, null=True)
+    price = models.FloatField(blank=True, null=True)
     # quantity = models.IntegerField(blank=True, null=True) #TODO - implement stock management
     
     def __str__(self) -> str:
